--- src/app/content/components/popUp/PrintButton.ts
@@ -1,6 +1,0 @@
-/**
- * PopUpPrintButton - Plain CSS implementation
- *
- * Migrated from styled-components to plain CSS.
- */
-export { default } from './PrintButton.impl';
